backend/requirements.txt
# PaiiD-2mx Backend - DEX Aggregator API
# Zero-cost architecture using free-tier APIs

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv==1.0.0

# HTTP & Web3
requests==2.31.0
httpx==0.25.2
web3==6.11.3

# Database (PostgreSQL)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
email-validator>=2.0.0
authlib>=1.3.0  # OAuth 2.0

# Error Tracking & Monitoring
sentry-sdk[fastapi]>=1.40.0

# Caching & Performance
redis>=5.0.0
cachetools>=5.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# CORS & Security
python-multipart>=0.0.6

# Rate Limiting
slowapi>=0.1.9
